disable bloom filters
